+++ src/pages/Account/components/ChangeNicknameModal.scss
@@ -53,16 +53,3 @@
         }
     }
 }
-
-.modal_background {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background: #000000bb;
-    z-index: 999;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
