+++ src/pages/Canvas/index.scss
@@ -23,7 +23,7 @@ a {
     .toolbox {
         display: flex;
         background-color: white;
-        border-radius: 10px;
+        border-radius: 8px;
         border: 1px solid #333;
         z-index: 100;
         margin: 10px auto 1em;
